gyro calib

+++ libraries/Arduino-MPU6050-master/MPU6050.h
@@ -24,6 +24,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #if ARDUINO >= 100
 #include "Arduino.h"
 #include <libmaple/i2c.h>
+#include <libmaple/iwdg.h>
+#include <HardwareTimer.h>
 #else
 #include "WProgram.h"
 #endif
@@ -233,7 +235,7 @@ class MPU6050
 	void setAccelOffsetZ(int16_t offset);
 
 	void calibrateGyro(uint8_t samples = 50);
-	void setThreshold(uint8_t multiple = 1);
+	void setThreshold(float multiple = 1);
 	uint8_t getThreshold(void);
 
 	SignedVector readRawGyro(void);
